requirements.txt
# Core Data & HTTP
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Embeddings & Vector Store
sentence-transformers>=2.3.1
annoy==1.17.3
chromadb==0.4.1

# LLM & Agents
langchain>=0.1.0
google-generativeai==0.3.2
langchain-google-genai>=0.0.6

# Editing & Validation
language-tool-python==2.7.1

# Utilities & Testing
python-dotenv==1.0.0
pytest==7.4.0
